+++ web/src/pages/upload_configure.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
-  import Link from '$/components/link.svelte'
-  import LoginButton from '$/components/login_button.svelte'
-  import Logo from '$/components/logo.svelte'
+  import Link from '$/components/Link.svelte'
+  import LoginButton from '$/components/LoginButton.svelte'
+  import Logo from '$/components/Logo.svelte'
   import { metatags } from '@roxi/routify'
   import { store as orange } from '$/components/blob/Orange.svelte'
   import { store as red } from '$/components/blob/Red.svelte'
@@ -9,7 +9,7 @@
   import { store as window } from '$/components/Window.svelte'
   import { fly } from 'svelte/transition'
   import { onMount } from 'svelte'
-  import Buton from '$/components/buton.svelte'
+  import Buton from '$/components/NavButton.svelte'
   let name = ''
   let isessay = true
   metatags.title = 'Eseuri'
